Add `pico_image.h` support
